Worked on printing floats  - Splitted the integer part from the decimal part of the float  - Printed both parts in console  - Pending : the checking and printing of the flags  - Pending : setting the precision flag on and checking if width is less than 6

+++ ft_putnbr_base.c
@@ -12,7 +12,7 @@
 
 #include "ft_printf.h"
 
-void	ft_putnbr_int(int n)
+void	ft_putnbr_int(long long n)
 {
 	if (n == -2147483648)
 	{
